+++ previous-versions/common/src/sdk/sdk.ts
@@ -60,8 +60,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.38.1";
-    genVersion = "2.131.1";
+    sdkVersion = "0.39.0";
+    genVersion = "2.139.1";
+    userAgent = "speakeasy-sdk/typescript 0.39.0 2.139.1 3.0.0 @codat/common";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);